+++ forui/lib/src/widgets/checkbox.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
+import 'package:forui/src/foundation/stateless_form_field.dart';
 
 import 'package:meta/meta.dart';
 
@@ -15,7 +16,7 @@ import 'package:forui/forui.dart';
 /// See:
 /// * https://forui.dev/docs/checkbox for working examples.
 /// * [FCheckboxStyle] for customizing a checkbox's appearance.
-class FCheckbox extends StatelessWidget {
+class FCheckbox extends FStatelessFormField<bool> {
   /// The style. Defaults to [FThemeData.checkboxStyle].
   final FCheckboxStyle? style;
 
@@ -42,44 +43,6 @@ class FCheckbox extends StatelessWidget {
   /// Called with true if this widget's node gains focus, and false if it loses focus.
   final ValueChanged<bool>? onFocusChange;
 
-  /// An optional method to call with the final value when the form is saved via [FormState.save].
-  final FormFieldSetter<bool>? onSave;
-
-  /// An optional method that validates an input. Returns an error string to display if the input is invalid, or null
-  /// otherwise.
-  ///
-  /// The returned value is exposed by the [FormFieldState.errorText] property.
-  final FormFieldValidator<bool>? validator;
-
-  /// An optional value to initialize the checkbox. Defaults to false.
-  final bool initialValue;
-
-  /// Whether the form is able to receive user input.
-  ///
-  /// Defaults to true. If [autovalidateMode] is not [AutovalidateMode.disabled], the checkbox will be auto validated.
-  /// Likewise, if this field is false, the widget will not be validated regardless of [autovalidateMode].
-  final bool enabled;
-
-  /// Used to enable/disable this checkbox auto validation and update its error text.
-  ///
-  /// Defaults to [AutovalidateMode.disabled].
-  ///
-  /// If [AutovalidateMode.onUserInteraction], this checkbox will only auto-validate after its content changes. If
-  /// [AutovalidateMode.always], it will auto-validate even without user interaction. If [AutovalidateMode.disabled],
-  /// auto-validation will be disabled.
-  final AutovalidateMode? autovalidateMode;
-
-  /// Restoration ID to save and restore the state of the checkbox.
-  ///
-  /// Setting the restoration ID to a non-null value results in whether or not the checkbox validation persists.
-  ///
-  /// The state of this widget is persisted in a [RestorationBucket] claimed from the surrounding [RestorationScope]
-  /// using the provided restoration ID.
-  ///
-  /// See also:
-  ///  * [RestorationManager], which explains how state restoration works in Flutter.
-  final String? restorationId;
-
   /// Creates a [FCheckbox].
   const FCheckbox({
     this.style,
@@ -90,88 +53,82 @@ class FCheckbox extends StatelessWidget {
     this.autofocus = false,
     this.focusNode,
     this.onFocusChange,
-    this.onSave,
-    this.validator,
-    this.initialValue = false,
-    this.enabled = true,
-    this.autovalidateMode,
-    this.restorationId,
+    super.onSave,
+    super.forceErrorText,
+    super.validator,
+    super.initialValue = false,
+    super.enabled = false,
+    super.autovalidateMode,
+    super.restorationId,
     super.key,
   });
 
   @override
-  Widget build(BuildContext context) {
+  Widget builder(BuildContext context, FormFieldState<bool> state) {
     final style = this.style ?? context.theme.checkboxStyle;
+    final stateStyle = switch ((enabled, state.hasError)) {
+      (true, false) => style.enabledStyle,
+      (false, false) => style.disabledStyle,
+      (_, true) => style.errorStyle,
+    };
+    final value = state.value ?? initialValue;
 
     return FocusableActionDetector(
+      enabled: enabled,
       autofocus: autofocus,
       focusNode: focusNode,
       onFocusChange: onFocusChange,
-      child: MouseRegion(
-        cursor: enabled ? SystemMouseCursors.click : MouseCursor.defer,
-        child: FormField<bool>(
-          onSaved: onSave,
-          validator: validator,
-          initialValue: initialValue,
-          enabled: enabled,
-          autovalidateMode: autovalidateMode,
-          restorationId: restorationId,
-          builder: (state) {
-            final stateStyle = enabled ? style.enabledStyle : style.disabledStyle;
-            final value = state.value ?? initialValue;
-            return Semantics(
-              label: semanticLabel,
-              enabled: enabled,
-              checked: value,
-              child: GestureDetector(
-                onTap: enabled
-                    ? () {
-                        state.didChange(!value);
-                        onChange?.call(!value);
-                      }
-                    : null,
-                child: FLabel(
-                  axis: Axis.horizontal,
-                  state: switch ((enabled, state.hasError)) {
-                    (true, false) => FLabelState.enabled,
-                    (false, false) => FLabelState.disabled,
-                    (_, true) => FLabelState.error,
-                  },
-                  label: label,
-                  description: description,
-                  error: Text(state.errorText ?? ''),
-                  child: AnimatedSwitcher(
-                    duration: style.animationDuration,
-                    switchInCurve: style.curve,
-                    child: SizedBox.square(
-                      key: ValueKey(value),
-                      dimension: 20,
-                      child: DecoratedBox(
-                        decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(4),
-                          border: Border.all(
-                            color: stateStyle.borderColor,
-                            width: 0.6,
-                          ),
-                          color: value ? stateStyle.checkedBackgroundColor : stateStyle.uncheckedBackgroundColor,
-                        ),
-                        child: value
-                            ? FAssets.icons.check(
-                                height: 15,
-                                width: 15,
-                                colorFilter: ColorFilter.mode(
-                                  stateStyle.iconColor,
-                                  BlendMode.srcIn,
-                                ),
-                              )
-                            : const SizedBox(),
-                      ),
+      mouseCursor: enabled ? SystemMouseCursors.click : MouseCursor.defer,
+      child: Semantics(
+        label: semanticLabel,
+        enabled: enabled,
+        checked: value,
+        child: GestureDetector(
+          onTap: enabled
+              ? () {
+            state.didChange(!value);
+            onChange?.call(!value);
+          }
+              : null,
+          child: FLabel(
+            axis: Axis.horizontal,
+            state: switch ((enabled, state.hasError)) {
+              (true, false) => FLabelState.enabled,
+              (false, false) => FLabelState.disabled,
+              (_, true) => FLabelState.error,
+            },
+            label: label,
+            description: description,
+            error: Text(state.errorText ?? ''),
+            child: AnimatedSwitcher(
+              duration: style.animationDuration,
+              switchInCurve: style.curve,
+              child: SizedBox.square(
+                key: ValueKey(value),
+                dimension: 16,
+                child: DecoratedBox(
+                  decoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(4),
+                    border: Border.all(
+                      color: stateStyle.borderColor,
+                      width: 0.6,
                     ),
+                    color: value ? stateStyle.checkedBackgroundColor : stateStyle.uncheckedBackgroundColor,
                   ),
+                  child: value
+                      ? FAssets.icons.check(
+                    height: 15,
+                    width: 15,
+                    colorFilter: ColorFilter.mode(
+                      stateStyle.iconColor,
+                      BlendMode.srcIn,
+                    ),
+                  )
+                      : const SizedBox(),
                 ),
               ),
-            );
-          },
+            ),
+          ),
         ),
       ),
     );
@@ -186,13 +143,7 @@ class FCheckbox extends StatelessWidget {
       ..add(ObjectFlagProperty.has('onChange', onChange))
       ..add(DiagnosticsProperty('autofocus', autofocus))
       ..add(DiagnosticsProperty('focusNode', focusNode))
-      ..add(ObjectFlagProperty.has('onFocusChange', onFocusChange))
-      ..add(ObjectFlagProperty.has('onSave', onSave))
-      ..add(ObjectFlagProperty.has('validator', validator))
-      ..add(DiagnosticsProperty('initialValue', initialValue))
-      ..add(DiagnosticsProperty('enabled', enabled))
-      ..add(EnumProperty('autovalidateMode', autovalidateMode))
-      ..add(StringProperty('restorationId', restorationId));
+      ..add(ObjectFlagProperty.has('onFocusChange', onFocusChange));
   }
 }
 
@@ -208,39 +159,50 @@ final class FCheckboxStyle with Diagnosticable {
   /// Defaults to [Curves.linear].
   final Curve curve;
 
+  /// The [FLabel]'s style.
+  final FLabelStyle? labelStyle;
+
   /// The checkbox's style when it's enabled.
   final FCheckboxStateStyle enabledStyle;
 
   /// The checkbox's style when it's disabled.
   final FCheckboxStateStyle disabledStyle;
 
+  /// The checkbox's style when it's in an error state.
+  final FCheckboxStateStyle errorStyle;
+
   /// Creates a [FCheckboxStyle].
   FCheckboxStyle({
     required this.enabledStyle,
     required this.disabledStyle,
+    required this.errorStyle,
     this.animationDuration = const Duration(milliseconds: 100),
     this.curve = Curves.linear,
+    this.labelStyle,
   });
 
   /// Creates a [FCheckboxStyle] that inherits its properties from the given [FColorScheme].
   FCheckboxStyle.inherit({required FColorScheme colorScheme, required FStyle style})
       : animationDuration = const Duration(milliseconds: 100),
         curve = Curves.linear,
+        labelStyle = FLabelStyles.inherit(style: style).horizontal,
         enabledStyle = FCheckboxStateStyle(
           borderColor: colorScheme.primary,
-          iconColor: colorScheme.background,
+          iconColor: colorScheme.primaryForeground,
           checkedBackgroundColor: colorScheme.primary,
           uncheckedBackgroundColor: colorScheme.background,
-          labelTextStyle: style.formFieldStyle.enabledStyle.labelTextStyle,
-          descriptionTextStyle: style.formFieldStyle.enabledStyle.descriptionTextStyle,
         ),
         disabledStyle = FCheckboxStateStyle(
           borderColor: colorScheme.primary.withOpacity(0.5),
-          iconColor: colorScheme.background.withOpacity(0.5),
+          iconColor: colorScheme.primaryForeground.withOpacity(0.5),
           checkedBackgroundColor: colorScheme.primary.withOpacity(0.5),
           uncheckedBackgroundColor: colorScheme.background.withOpacity(0.5),
-          labelTextStyle: style.formFieldStyle.disabledStyle.labelTextStyle,
-          descriptionTextStyle: style.formFieldStyle.disabledStyle.descriptionTextStyle,
+        ),
+        errorStyle = FCheckboxStateStyle(
+          borderColor: colorScheme.error,
+          iconColor: colorScheme.errorForeground,
+          checkedBackgroundColor: colorScheme.error,
+          uncheckedBackgroundColor: colorScheme.background,
         );
 
   /// Returns a copy of this [FCheckboxStyle] with the given properties replaced.
@@ -263,14 +225,18 @@ final class FCheckboxStyle with Diagnosticable {
   FCheckboxStyle copyWith({
     Duration? animationDuration,
     Curve? curve,
+    FLabelStyle? labelStyle,
     FCheckboxStateStyle? enabledStyle,
     FCheckboxStateStyle? disabledStyle,
+    FCheckboxStateStyle? errorStyle,
   }) =>
       FCheckboxStyle(
         animationDuration: animationDuration ?? this.animationDuration,
         curve: curve ?? this.curve,
+        labelStyle: labelStyle ?? this.labelStyle,
         enabledStyle: enabledStyle ?? this.enabledStyle,
         disabledStyle: disabledStyle ?? this.disabledStyle,
+        errorStyle: errorStyle ?? this.errorStyle,
       );
 
   @override
@@ -279,8 +245,10 @@ final class FCheckboxStyle with Diagnosticable {
     properties
       ..add(DiagnosticsProperty('animationDuration', animationDuration))
       ..add(DiagnosticsProperty('curve', curve))
+      ..add(DiagnosticsProperty('labelStyle', labelStyle))
       ..add(DiagnosticsProperty('enabledStyle', enabledStyle))
-      ..add(DiagnosticsProperty('disabledStyle', disabledStyle));
+      ..add(DiagnosticsProperty('disabledStyle', disabledStyle))
+      ..add(DiagnosticsProperty('errorStyle', errorStyle));
   }
 
   @override
@@ -290,11 +258,19 @@ final class FCheckboxStyle with Diagnosticable {
           runtimeType == other.runtimeType &&
           animationDuration == other.animationDuration &&
           curve == other.curve &&
+          labelStyle == other.labelStyle &&
           enabledStyle == other.enabledStyle &&
-          disabledStyle == other.disabledStyle;
+          disabledStyle == other.disabledStyle &&
+          errorStyle == other.errorStyle;
 
   @override
-  int get hashCode => animationDuration.hashCode ^ curve.hashCode ^ enabledStyle.hashCode ^ disabledStyle.hashCode;
+  int get hashCode =>
+      animationDuration.hashCode ^
+      curve.hashCode ^
+      labelStyle.hashCode ^
+      enabledStyle.hashCode ^
+      disabledStyle.hashCode ^
+      errorStyle.hashCode;
 }
 
 /// A checkbox state's style.
@@ -311,20 +287,12 @@ final class FCheckboxStateStyle with Diagnosticable {
   /// The unchecked checkbox's background color.
   final Color uncheckedBackgroundColor;
 
-  /// The label's text style.
-  final TextStyle labelTextStyle;
-
-  /// The description's text style.
-  final TextStyle descriptionTextStyle;
-
   /// Creates a [FCheckboxStateStyle].
   const FCheckboxStateStyle({
     required this.borderColor,
     required this.iconColor,
     required this.checkedBackgroundColor,
     required this.uncheckedBackgroundColor,
-    required this.labelTextStyle,
-    required this.descriptionTextStyle,
   });
 
   /// Returns a copy of this [FCheckboxStateStyle] with the given properties replaced.
@@ -357,8 +325,6 @@ final class FCheckboxStateStyle with Diagnosticable {
         iconColor: iconColor ?? this.iconColor,
         checkedBackgroundColor: checkedBackgroundColor ?? this.checkedBackgroundColor,
         uncheckedBackgroundColor: uncheckedBackgroundColor ?? this.uncheckedBackgroundColor,
-        labelTextStyle: labelTextStyle ?? this.labelTextStyle,
-        descriptionTextStyle: descriptionTextStyle ?? this.descriptionTextStyle,
       );
 
   @override
@@ -368,9 +334,7 @@ final class FCheckboxStateStyle with Diagnosticable {
       ..add(ColorProperty('borderColor', borderColor))
       ..add(ColorProperty('checkedIconColor', iconColor))
       ..add(ColorProperty('checkedBackgroundColor', checkedBackgroundColor))
-      ..add(ColorProperty('uncheckedBackgroundColor', uncheckedBackgroundColor))
-      ..add(DiagnosticsProperty('labelTextStyle', labelTextStyle))
-      ..add(DiagnosticsProperty('descriptionTextStyle', descriptionTextStyle));
+      ..add(ColorProperty('uncheckedBackgroundColor', uncheckedBackgroundColor));
   }
 
   @override
@@ -381,16 +345,12 @@ final class FCheckboxStateStyle with Diagnosticable {
           borderColor == other.borderColor &&
           iconColor == other.iconColor &&
           checkedBackgroundColor == other.checkedBackgroundColor &&
-          uncheckedBackgroundColor == other.uncheckedBackgroundColor &&
-          labelTextStyle == other.labelTextStyle &&
-          descriptionTextStyle == other.descriptionTextStyle;
+          uncheckedBackgroundColor == other.uncheckedBackgroundColor;
 
   @override
   int get hashCode =>
       borderColor.hashCode ^
       iconColor.hashCode ^
       checkedBackgroundColor.hashCode ^
-      uncheckedBackgroundColor.hashCode ^
-      labelTextStyle.hashCode ^
-      descriptionTextStyle.hashCode;
+      uncheckedBackgroundColor.hashCode;
 }